Create LanguageController and use it in switcher directive and in user details page
