+++ client/views/layout/navbar.html
@@ -14,7 +14,7 @@
 
     <nav class="navbar-collapse" uib-collapse="!navbar.collapsed">
       <ul class="nav navbar-nav navbar-right">
-        <li navbar-language-switcher></li>
+        <li language-switcher></li>
         <li ng-if="'home' | isState">
           <a href ui-sref="home" sb-scroll-to="[name=forms]" translate>NAV_FORMS</a>
         </li>
