+++ client/views/login.html
@@ -49,7 +49,7 @@
         </div>
         <div class="panel-footer">
           <p>
-            <a class="language" navbar-language-switcher></a>
+            <a class="language" language-switcher></a>
             <a ui-sref="forgot" href translate>LOGIN_FORGOT_PASSWORD</a></p>
           <p>
             <!--<p><a ui-sref="register" href translate>LOGIN_REGISTER</a></p>-->
