+++ client/views/register.html
@@ -67,7 +67,7 @@
         <div class="panel-footer">
           <div class="form-actions">
             <div class="btn-group pull-right btn-group-lg" >
-            <a class="language" navbar-language-switcher></a>
+            <a class="language" language-switcher></a>
             </div>
             <button type="submit"
                     ng-disabled="form.$invalid || sessionController.loading || !sessionController.tos_accepted || !user.gdprConsent"
